Butler

Description

Welcome to Butler.

This module exposes the main API.

Core primitive

runMain :: IO a -> IO a

Helper to setup line-buffering and cgroup rts for container

data ProcessIO a

spawnInitProcess :: RawFilePath -> ProcessIO a -> IO ExitReason

Run the initial process with a given storage root directory.

Concurrency API

data Process

A process is a killable thread.

Constructors

Process

Fields

pid :: Pid

the process id

childs :: TVar [Process]

the list of sub process

scope :: Scope

the processor scope

threadId :: ThreadId

the haskell rts id

thread :: Thread ExitReason

the ki thread reference

doneVar :: TMVar ()

a sync var to terminate the process ki scope

status :: TVar ProcessStatus

the process status

program :: ProgramName

createdAt :: Time

newtype ProgramName

Constructors

ProgramName Text

spawnProcess :: ProgramName -> ProcessIO () -> ProcessIO Process

Create a process.

waitProcess :: MonadIO m => Process -> m ExitReason

stopProcess :: Process -> STM Bool

Stop a process, returns False if the process did not received the message.

killProcess :: Pid -> ProcessIO Bool

Lookup and kill the pid. Use stopProcess instead.

spawnThread_ :: ProcessIO Void -> ProcessIO ()

Create a child thread that never exit, for example to handle messages. If the thread crash, the whole process is terminated.

spawnThread :: ProcessIO a -> ProcessIO (Thread a)

Create a child thread.

module Butler.Core.Pipe

Logging API

logInfo :: HasCallStack => Text -> [Pair] -> ProcessIO ()

Use logInfo for nominal but important event.

logError :: HasCallStack => Text -> [Pair] -> ProcessIO ()

Use logError for unexpected mal-functions.

logDebug :: HasCallStack => Text -> [Pair] -> ProcessIO ()

Use logDebug for the rest.

Time API

data Time

The number of milli secondes since the clock started.

getTime :: ProcessIO Time

data Milli

A duration in millisecond

sleep :: MonadIO m => Milli -> m ()

Pause the execution.

data WaitResult result

Constructors

WaitCompleted result

WaitTimeout

waitTransaction :: MonadIO m => Milli -> STM result -> m (STM (WaitResult result))

Wait for a transaction with a timeout.

Network API

module Butler.Core.Network

Storage API

newProcessMemory :: Serialise a => StorageAddress -> ProcessIO a -> ProcessIO (a, MemoryVar a)

module Butler.Core.Memory

newtype StorageAddress

Constructors

StorageAddress ByteString

getPath :: StorageAddress -> ProcessIO RawFilePath

module Butler.Core.File

Display API

module Butler.Display.Client

data DisplayEvent

Constructors

UserConnected ChannelName DisplayClient

UserDisconnected ChannelName DisplayClient

App API

data App

A graphical application definition.

Constructors

App

Fields

name :: ProgramName

The application name.

tags :: Set AppTag

Its categories.

title :: Text

Its title.

description :: Text

A description.

size :: Maybe (Int, Int)

An optional size.

xfiles :: [XStaticFile]

Required XStaticFile

acceptFiles :: Maybe ContentType

start :: AppContext -> ProcessIO ()

Start action.

data AppSharedContext

Constructors

AppSharedContext

Fields

display :: Display

processEnv :: ProcessEnv

appSet :: AppSet

dynamics :: Dynamics

apps :: Apps

data AppContext

The application context

Constructors

AppContext

Fields

clients :: DisplayClients

the list of all the connected clients. To send update, app should uses `sendsHtml clients ""`

wid :: WinID

the instance identifier. The app should mount its UI with `with div_ [wid_ wid] "body"`, and the trigger must container the WinID suffix too.

pipe :: Pipe AppEvent

the channel to receive events.

shared :: AppSharedContext

defaultApp :: ProgramName -> (AppContext -> ProcessIO ()) -> App

data AppEvent

The type of event an app receive

Constructors

AppDisplay DisplayEvent
A display event (e.g. to mount the UI)

AppTrigger GuiEvent
A trigger event (e.g. onclick)

AppData DataEvent
A data event (e.g. for raw data)

AppFile Directory (Maybe File)
A file event (e.g. a new file opened)

data GuiEvent

Constructors

GuiEvent

Fields

client :: DisplayClient

trigger :: TriggerName

body :: Value

data DataEvent

Constructors

DataEvent

Fields

client :: DisplayClient

buffer :: ByteString

The data buffer (without the channel id)

rawBuffer :: ByteString

The original buffer including the channel id. Use this to forward the message to other clients.

serveApps :: DisplayApplication -> [App] -> ProcessIO Void

Serve applications with one instance per client.

serveDashboardApps :: DisplayApplication -> [App] -> ProcessIO Void

Serve applications with one instance for all clients.

publicDisplayApp :: PageTitle -> Maybe PageDesc -> DisplayApplication

newtype Service

Constructors

Service App

GUI toolkit

sendHtmlOnConnect :: HtmlT STM () -> AppEvent -> ProcessIO ()

A convenient helper to mount the UI when a new user connect.

module Butler.Display.GUI

module Butler.Window

Prelude

module Butler.Prelude