Butler.Core

Boot

data OS

Constructors

OS

Fields

processor :: Processor

storage :: Storage

clock :: Clock

logger :: Logger SystemEvent

buzzer :: Buzzer

spawnInitProcess :: RawFilePath -> ProcessIO a -> IO ExitReason

Run the initial process with a given storage root directory.

withButlerOS :: ProcessIO a -> IO ExitReason

data ProcessEnv

Constructors

ProcessEnv

Fields

os :: OS

process :: Process

data ProcessIO a

runProcessIO :: OS -> Process -> ProcessIO a -> IO a

getSelfProcess :: ProcessIO Process

asProcess :: ProcessEnv -> ProcessIO a -> ProcessIO a

runExternalProcess :: Text -> ProcessConfig stdin stdout0 stderr0 -> ProcessIO ()

Memory api

getPath :: StorageAddress -> ProcessIO RawFilePath

newProcessMemory :: Serialise a => StorageAddress -> ProcessIO a -> ProcessIO (a, MemoryVar a)

chroot :: StorageAddress -> ProcessIO a -> ProcessIO a

Change the root storage directory of a process.

Processor api

spawnProcess :: ProgramName -> ProcessIO () -> ProcessIO Process

Create a process.

superviseProcess :: ProgramName -> ProcessIO Void -> ProcessIO Process

stopProcess :: Process -> STM Bool

Stop a process, returns False if the process did not received the message.

killProcess :: Pid -> ProcessIO Bool

Lookup and kill the pid. Use stopProcess instead.

spawnThread_ :: ProcessIO Void -> ProcessIO ()

Create a child thread that never exit, for example to handle messages. If the thread crash, the whole process is terminated.

spawnThread :: ProcessIO a -> ProcessIO (Thread a)

Create a child thread.

Log api

logSystem :: EventSeverity -> SystemEvent -> ProcessIO ()

logDebug :: HasCallStack => Text -> [Pair] -> ProcessIO ()

Use logDebug for the rest.

logInfo :: HasCallStack => Text -> [Pair] -> ProcessIO ()

Use logInfo for nominal but important event.

logError :: HasCallStack => Text -> [Pair] -> ProcessIO ()

Use logError for unexpected mal-functions.

Clock api

getTime :: ProcessIO Time

IPC api

writePipe :: HasCallStack => Pipe a -> a -> ProcessIO ()

Helpers

waitProcess :: MonadIO m => Process -> m ExitReason

Re-exports

newtype ProgramName

Constructors

ProgramName Text

data Process

A process is a killable thread.

Constructors

Process

Fields

pid :: Pid

the process id

childs :: TVar [Process]

the list of sub process

scope :: Scope

the processor scope

threadId :: ThreadId

the haskell rts id

thread :: Thread ExitReason

the ki thread reference

doneVar :: TMVar ()

a sync var to terminate the process ki scope

status :: TVar ProcessStatus

the process status

program :: ProgramName

createdAt :: Time

newtype ProcessAction

Constructors

ProcessAction (Process -> IO ())

data SystemEvent

Constructors

ProcessCreated Process

ProcessStopped Process ExitReason

ProcessMessage ByteString Process Text

StorageSync Int

SystemReady

SystemCompleted

DaemonCrashed Process Int