Butler.Core.File

Description

This module contains the file system logic.

File system data types

data Entry

Entry is the base element of a file system.

Constructors

File File

Directory Directory

data File

data Directory

readRootDirectory :: MonadIO m => RawFilePath -> m (Maybe Directory)

Read the root directory content, return Nothing when the file path is not a directory. You must call this function to get the first directory.

Directory API

newDirectory :: Directory -> FileName -> STM (Maybe Directory)

getRootDir :: Directory -> Directory

Return the outer most directory.

data FileName

FileName are provided by the user, they must not begin or end with a /.

lookupChild :: Directory -> FileName -> STM (Maybe Entry)

baseName :: RawFilePath -> RawFilePath

deleteDirectory :: MonadIO m => Directory -> Directory -> m ()

File API

createFile :: MonadIO m => Directory -> FileName -> ByteString -> m File

readFileBS :: MonadIO m => Directory -> File -> m ByteString

writeFileBS :: MonadIO m => Directory -> File -> ByteString -> m ()

deleteFile :: MonadIO m => Directory -> File -> m ()

Rename/Move API

renameEntry :: MonadIO m => Directory -> Entry -> FileName -> m ()

moveEntry :: MonadIO m => Directory -> Directory -> Entry -> m ()

External API

data FileLoc

The absolute location of a directory, for remote user.

getFileLoc :: Directory -> Maybe File -> FileLoc

Get the absolute location of a directory.

resolveFileLoc :: Directory -> FileLoc -> STM (Maybe (Directory, Maybe File))

Upload API

data PartialFile

createPartialFile :: MonadIO m => Directory -> FileName -> m PartialFile

appendPartialFile :: MonadIO m => PartialFile -> ByteString -> m ()

finalizePartialFile :: MonadIO m => PartialFile -> FileOffset -> m File

Mime API

data ContentType

Constructors

BinaryContent

ImageContent

VideoContent

TextContent

ArchiveContent

fileContentType :: File -> ContentType