Butler.Core.Processor

Documentation

data Processor

withProcessor :: (Processor -> IO a) -> IO a

awaitProcessor :: Processor -> STM ()

data Process

A process is a killable thread.

Constructors

Process

Fields

pid :: Pid

the process id

childs :: TVar [Process]

the list of sub process

scope :: Scope

the processor scope

threadId :: ThreadId

the haskell rts id

thread :: Thread ExitReason

the ki thread reference

doneVar :: TMVar ()

a sync var to terminate the process ki scope

status :: TVar ProcessStatus

the process status

program :: ProgramName

createdAt :: Time

newtype ProcessAction

Constructors

ProcessAction (Process -> IO ())

lookupChildProcess :: Process -> Pid -> STM (Maybe Process)

lookupProcess :: Processor -> Pid -> STM (Maybe Process)

startProcess :: Clock -> Logger SystemEvent -> Processor -> Maybe Process -> ProgramName -> ProcessAction -> IO Process

stopProcess :: Process -> STM Bool

Stop a process, returns False if the process did not received the message.

useful re-exports

module Butler.Core.Process