Butler.Database

Description

This module contains logic around sqlite-easy

Documentation

data Database

withDatabase :: StorageAddress -> DatabaseMigration -> (Database -> ProcessIO a) -> ProcessIO a

data DatabaseMigration

Constructors

DatabaseMigration

Fields

migrations :: [MigrationName]

migrateUp :: MigrationName -> Database -> ProcessIO ()

migrateDown :: MigrationName -> Database -> ProcessIO ()

dbSimpleCreate :: Text -> Text -> DatabaseMigration

dbExecute :: MonadUnliftIO m => Database -> Query -> [NamedParam] -> m ()

dbQuery :: (MonadUnliftIO m, FromRow r) => Database -> Query -> [NamedParam] -> m [r]

re-export

data NamedParam where

Constructors

(:=) :: forall v. ToField v => Text -> v -> NamedParam infixr 3

newtype Only a

The 1-tuple type or single-value "collection".

This type is structurally equivalent to the Identity type, but its intent is more about serving as the anonymous 1-tuple type missing from Haskell for attaching typeclass instances.

Parameter usage example:

encodeSomething (Only (42::Int))

Result usage example:

xs <- decodeSomething
forM_ xs $ \(Only id) -> {- ... -}

Constructors

Only

Fields

fromOnly :: a