Butler.Display.Client

Description

This module contains the DisplayClient logic.

Display client

data DisplayClient

A network client. Note that the sendThread must be running for the send* function to work.

Constructors

DisplayClient

Fields

conn :: Connection

endpoint :: Endpoint

process :: Process

session :: Session

recv :: TVar Word64

send :: TVar Word64

sendChannel :: TChan DataMessage

Html output

sendHtml :: DisplayClient -> HtmlT STM () -> STM ()

Send Html to a client.

sendsHtml :: DisplayClients -> HtmlT STM () -> ProcessIO ()

Send Html to all clients.

sendsHtmlButSelf :: DisplayClient -> DisplayClients -> HtmlT STM () -> ProcessIO ()

Send Html to all clients except the provided one (self).

clientsDraw :: DisplayClients -> (DisplayClient -> ProcessIO (HtmlT STM ())) -> ProcessIO ()

Send html using a draw action.

Binary output

sendBinary :: DisplayClient -> LByteString -> STM ()

Send binary to a client.

sendsBinary :: DisplayClients -> LByteString -> ProcessIO ()

Send binary to all clients.

sendsBinaryButSelf :: DisplayClient -> DisplayClients -> LByteString -> ProcessIO ()

Send binary to all clients except the provided one (self).

Inputs

recvData :: MonadIO m => DisplayClient -> m DataMessage

Low-level helper to read a DataMessage.

recvBinary :: MonadIO m => DisplayClient -> m ByteString

Low-level helper to read a binary buffer.

Management thread

pingThread :: DisplayClient -> ProcessIO Void

An action to keep a client alive.

sendThread :: DisplayClient -> ProcessIO Void

An action to send data message.

Collection of clients

data DisplayClients

A collection of clients.

newDisplayClients :: STM DisplayClients

getClients :: DisplayClients -> STM [DisplayClient]

addClient :: DisplayClients -> DisplayClient -> STM ()

delClient :: DisplayClients -> DisplayClient -> STM ()

Internal

newClient :: Connection -> Endpoint -> Process -> Session -> STM DisplayClient

newtype Endpoint

Constructors

Endpoint Text