Butler.Frame

Description

Exchange format for API client. A frame contains html elements or binary data such as terminal output.

Documentation

data DataEvent

Constructors

DataEvent

Fields

client :: DisplayClient

buffer :: ByteString

The data buffer (without the channel id)

rawBuffer :: ByteString

The original buffer including the channel id. Use this to forward the message to other clients.

helper

encodeMessageL :: WinID -> LByteString -> LByteString

encodeMessage :: Natural -> LByteString -> LByteString

decodeMessage :: ByteString -> Maybe (Natural, ByteString)

butlerHelpersScript :: Text