Butler.Service.FileService

Documentation

data VolumeName

VolumeName is the name of a root directory, e.g. Documents or Desktop

getVolumeDirectory :: AppSharedContext -> Maybe VolumeName -> ProcessIO Directory

getVolumeDirectory enables app to get access to the File API.

fileService :: Service

filesUploadButton :: Monad m => WinID -> FileLoc -> HtmlT m ()